Modularize todo api
